+++ app/templates/signin.html
@@ -17,7 +17,7 @@
 
 
         <!-- Login Form -->
-        <form id="login-form" hx-post="http://127.0.0.1:8000/user/login"
+        <form id="login-form" hx-post="http://127.0.0.1:8000/auth/user/login"
             hx-headers='{"Content-Type": "application/json"}' hx-trigger="submit" hx-ext="json-enc"
             class="box">
             <div class="field">
