+++ app/templates/signup.html
@@ -15,7 +15,7 @@
         <h1 class="title has-text-centered mt-5">Signup</h1>
 
         <!-- Signup Form -->
-        <form id="signup-form" hx-post="http://127.0.0.1:8000/user/create"
+        <form id="signup-form" hx-post="http://127.0.0.1:8000/auth/user/create"
             hx-headers='{"Content-Type": "application/json"}' hx-trigger="submit" hx-ext="json-enc"
             class="box">
             <div class="field">
